ColorPyAPI/requirement.txt
blinker==1.7.0
boto3==1.34.64
botocore==1.34.64
click==8.1.7
# Editable install with no version control (decompose==0.1.0)
-e ./decompose-main
defusedxml==0.7.1
filelock==3.13.1
Flask==3.0.2
fsspec==2024.3.0
guided-filter-pytorch==3.7.5
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
Pillow==10.1.0
pyora==0.3.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sympy==1.12
torch==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
gunicorn==19.9.0
